Final Hackathon/requirements.txt
streamlit==1.34.0
python-dotenv==1.0.1
langchain==0.1.20
langgraph==0.0.30
langchain-community==0.0.38
langchain-google-genai==0.0.7
langsmith==0.1.81

sentence-transformers==2.7.0
chromadb==0.4.7
chroma-hnswlib==0.7.2

pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0

requests==2.31.0
beautifulsoup4==4.12.2
fastapi==0.99.1
starlette==0.27.0

protobuf==4.25.3
